Add preconnect header for unpkg.com
Cloudflare Pages uses the `_headers` file
to declare custom headers for resources.
In this case, we want requests to the document root
to return a `Link` header that hints to the browser
that it should connect to `https://unpkg.com`.
We also include headers that Cloudflare Pages
would normally include on its own,
but since we are providing custom headers,
they are no longer included by default.